feat(fedimint-cli): add setup code decoding functionality
